backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["vaduska <vadimsalapugin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
python-dotenv = "^0.21.1"
flake8 = "^6.0.0"
flake8-bugbear = "23.1.14"
flake8-builtins = "2.1.0"
pep8-naming = "0.13.3"
flake8-commas = "2.1.0"
flake8-variables-names = "0.0.5"
flake8-import-order = "0.18.2"
types-requests = "2.28.11.7"
autoflake = "^2.0.0"
black = "^22.12.0"
psycopg2-binary = "2.9.5"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.13.0"
pyjwt = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
